pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ndlocr-lite-adapter"
version = "0.1.0"
description = "Immich machine-learning compatible OCR adapter backed by NDLOCR-Lite."
readme = "README.md"
requires-python = ">=3.11,<4.0"
license = "MIT"
dependencies = [
  "fastapi>=0.116.1,<1.0",
  "orjson>=3.11.3,<4.0",
  "pillow>=11.3.0,<13.0",
  "pydantic-settings>=2.10.1,<3.0",
  "python-multipart>=0.0.20,<1.0",
  "uvicorn[standard]>=0.35.0,<1.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1,<1.0",
  "pytest>=8.4.2,<9.0",
  "ruff>=0.13.0,<1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/ndlocr_lite_adapter"]

[dependency-groups]
dev = [
  "httpx>=0.28.1,<1.0",
  "pytest>=8.4.2,<9.0",
  "ruff>=0.13.0,<1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
